Added Toon Fantasy Nature terrain assets.
Added terrain assets from the pack. These were originally excluded due to the size of the assets, but now we're starting to use them in the project, so let's commit them to the repo.

Note that the previous commit for the demo scene made use of these uncommitted assets! This commit should fix the broken scene.