Forced LOD level 8 in quality settings.
This is a bit of a hack/test. The third-party assets have no LOD fade so we see really ugly popping (especially on the grass) when running around. Setting level 8 (i.e. max LOD level) ensures we always render the highest quality LOD for all objects, which sort of makes sense because we're using an isometric perspective. So for now, we're at max quality all the time. If there are performance issues down the line, we'll adjust.